+++ Eyes.Sdk.Core.DotNet/IEyes.cs
@@ -1,5 +1,4 @@
-﻿using Applitools.VisualGrid;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace Applitools
 {