+++ Eyes.Sdk.Core.DotNet/RunningTest.cs
@@ -161,8 +161,6 @@ namespace Applitools
             IssueAbort(e, true);
         }
 
-        public string TestId { get; private set; } = Guid.NewGuid().ToString();
-
         public TestResultContainer GetTestResultContainer() { return testResultContainer_; }
     }
 }